requirements.txt
torch>=2.8.0
torch-geometric>=2.3.0
pymatgen>=2023.5.31
mp-api>=0.37.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0